+++ examples/pingpong/Session.cpp
@@ -30,7 +30,7 @@ void Session::Close()
 {
     if (id_ > 0)
     {
-        fprintf(stderr, "%d closed\n", fd_);
+        fprintf(stderr, "%d closed\n", (int)fd_);
         closesocket(fd_);
         fd_ = INVALID_SOCKET;
         int sid = id_;
@@ -60,7 +60,7 @@ int Session::Write(const void* data, int len)
 void Session::StartRead()
 {
     recv_buf_.resize(1024);
-    recv_ctx_->SetBuffer(&recv_buf_[0], recv_buf_.size());
+    recv_ctx_->SetBuffer(&recv_buf_[0], (int)recv_buf_.size());
     recv_ctx_->cb = std::bind(&Session::OnRead, this, recv_ctx_);
     service_->AsyncRead(recv_ctx_);
 }
@@ -80,7 +80,7 @@ void Session::OnRead(OverlapContext* ctx)
     int nbytes = ctx->GetTransferredBytes();
     if (nbytes > 0)
     {
-        fprintf(stdout, "%d recv %d bytes, %d\n", fd_, nbytes, recv_count_);
+        fprintf(stdout, "%d recv %d bytes, %d\n", (int)fd_, nbytes, recv_count_);
         recv_buf_.resize(nbytes);
         recv_count_++;
         Write(&recv_buf_[0], nbytes);
@@ -101,7 +101,7 @@ void Session::OnWritten(OverlapContext* ctx)
     int nbytes = ctx->GetTransferredBytes();
     if (nbytes > 0)
     {
-        fprintf(stdout, "%d send %d bytes, %d\n", fd_, nbytes, recv_count_);
+        fprintf(stdout, "%d send %d bytes, %d\n", (int)fd_, nbytes, recv_count_);
     }
 
     delete ctx->buf.buf;
